Move calculateDistance to utils
